feat: Add "Timer" module to repo - allows a timer to be set and repeat a given function for a duration and then stop repeating the function.
Note: The functions can also be stopped before they start, and stopped
while they are running using the output array. The first element stops
the initial timer until the functions start playing and repeating, and
the second element stops the repeating function once it has begun
(given as a Promise). Got the code working with no global variables,
which was a little painful but overall a nice accomplishment!

Note PS: Need to figure out how to import modules into the browser -
apparently, "require" does not work and a plugin is needed...